fix(onboarding): wrap useSearchParams in Suspense boundary
Fix Next.js 15 build error by separating onboarding logic into a
separate component and wrapping it in Suspense.

Error: useSearchParams() should be wrapped in a suspense boundary at
page "/onboarding"

Changes:
- Extract onboarding logic to onboarding-content.tsx client component
- Wrap OnboardingContent in Suspense with loading fallback
- Satisfies Next.js 15 requirement for useSearchParams()

Fixes Vercel build failure.
